Move Catalog and Schema to catalog_schema module
